Replace the Gradle/AsciidoctorJ build with plain Asciidoctor
The Gradle-based build had multiple problems and quirks: insufficient
Java heap space when generating PDFs, bugs in the version selection for
Ruby gems, and now a JCenter outage that made the build fail completely
for a short time because some plugin dependencies couldn't be resolved.

The build is now simple enough that a plain Makefile does the job as
well, using the regular command-line version of Asciidoctor. This is
noticeably faster than the Gradle build and hopefully should be simpler
to maintain as well.